requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
fastapi==0.73.0
torchvision==0.9.0+cpu
torch==1.8.0+cpu
python-multipart==0.0.5
Pillow==9.0.0
uvicorn==0.17.0.post1
gunicorn==20.1.0
pymongo==4.1.0
